Convert to webpack
Parcel and wasm-bindgen do not work well together yet.

Rather than having to run three separate start targets, this runs one
integrated configuration, which I find somewhat faster and more flexible
to work with, but otherwise is like for like. It should be possible to
take advantage of webpack features (e.g. UMD library support) to further
improve things, but this seems sufficient for now to enable the
performance work we're focused on.