Return to using PUBLISHER_ADDR
We need to set the scheme and port in the publisher endpoint, so an
Address is a better conceptual name for what we're doing than a Host.

I've also made all the demos work in theory the same way, though until
we have this running I can't be quite sure that its all working
together.